+++ AndroidSource/src/Components/AwakeManager.h
@@ -25,12 +25,12 @@ public:
 	void Activate(class Application* app);
 	void Deactivate(class Application* app) {};
 	void Update(class Application* app);
-	shared::ServerResponse KeepAwake(class Application* app, uint32_t duration_ms);
+	shared::ServerResponse KeepAwake(class Application* app, long duration_ms);
 
 	bool IsScreenActive() { return awake_state != AwakeState::Update; };
 
 private:
-	shared::ServerResponse KeepAwake_Internal(class Application* app, uint32_t duration_ms);
+	shared::ServerResponse KeepAwake_Internal(class Application* app, long duration_ms);
 	shared::ServerResponse TurnScreenOn(struct android_app* state, long duration);
 	AwakeState awake_state = AwakeState::Idle;
 	SimpleTimer awake_timer;
